+++ test/integration/provider-scenarios/daemon-http-server.test.ts
@@ -12,7 +12,7 @@ import {
   trackDownloadableArtifact,
 } from '../../../src/daemon/artifact-tracking.ts';
 import { DAEMON_RPC_PROTOCOL_VERSION } from '../../../src/daemon/http-health.ts';
-import { createDaemonHttpServer } from '../../../src/daemon/http-server.ts';
+import { createDaemonHttpServer } from '../../../src/daemon/server/http-server.ts';
 import { emitRequestProgress } from '../../../src/daemon/request-progress.ts';
 import { getRequestSignal, isRequestCanceled } from '../../../src/daemon/request-cancel.ts';
 import type { DaemonRequest, DaemonResponse } from '../../../src/daemon/types.ts';
